requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
aws==0.2.5
bcrypt==4.2.1
beautifulsoup4==4.12.3
boto==2.49.0
boto3==1.35.65
botocore==1.35.65
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
cryptography==43.0.3
decorator==5.1.1
Deprecated==1.2.15
distro==1.9.0
Django==5.0.7
django-bootstrap4==24.4
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-taggit==6.1.0
djangorestframework==3.15.2
dnspython==2.6.1
fabric==3.2.2
flatbuffers==24.3.25
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
humanfriendly==10.0
idna==3.10
imageio==2.35.1
instamojo-wrapper==1.2.0
invoke==2.2.0
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
llvmlite==0.43.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
openai==1.52.2
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
paramiko==3.5.0
paypalrestsdk==1.13.3
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
prettytable==3.12.0
propcache==0.2.0
protobuf==5.28.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMatting==1.1.12
pymongo==4.8.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2024.1
razorpay==1.4.2
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
rpds-py==0.20.0
s3transfer==0.10.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
sslcommerz-lib==1.0
stripe==11.1.1
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2024.9.20
tqdm==4.66.5
twilio==9.3.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.16.0
